+++ experimental/CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to experimental packages in this project will be documented 
 
 ### :bug: (Bug Fix)
 
+* fix: Hide `require.resolve` from bundlers
+
 ### :books: (Refine Doc)
 
 ### :house: (Internal)